Move KafkaConsumer to common utils package
